+++ core/pom.xml
@@ -47,7 +47,6 @@
         <module>endpoints-spi</module>
         <module>imds</module>
         <module>retries-api</module>
-        <module>retries</module>
     </modules>
 
     <dependencyManagement>
