+++ core/retries-api/src/main/java/software/amazon/awssdk/retries/api/internal/backoff/FixedDelayWithoutJitter.java
@@ -13,26 +13,35 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.retries.backoff;
+package software.amazon.awssdk.retries.api.internal.backoff;
 
 import java.time.Duration;
 import software.amazon.awssdk.annotations.SdkInternalApi;
 import software.amazon.awssdk.retries.api.BackoffStrategy;
+import software.amazon.awssdk.utils.ToString;
 import software.amazon.awssdk.utils.Validate;
 
 /**
- * Strategy that do not back off: retry immediately.
+ * Strategy that waits for a period of time equal to the provided delay.
  */
 @SdkInternalApi
-final class Immediately implements BackoffStrategy {
+public final class FixedDelayWithoutJitter implements BackoffStrategy {
+    private final Duration delay;
+
+    public FixedDelayWithoutJitter(Duration delay) {
+        this.delay = Validate.isPositive(delay, "delay");
+    }
+
     @Override
     public Duration computeDelay(int attempt) {
         Validate.isPositive(attempt, "attempt");
-        return Duration.ZERO;
+        return delay;
     }
 
     @Override
     public String toString() {
-        return "(Immediately)";
+        return ToString.builder("FixedDelayWithoutJitter")
+                       .add("delay", delay)
+                       .build();
     }
 }
